normed, needs -R flags

+++ sources/check_ls.c
@@ -6,35 +6,35 @@
 /*   By: asultanb <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/18 16:42:07 by asultanb          #+#    #+#             */
-/*   Updated: 2020/01/21 16:21:50 by asultanb         ###   ########.fr       */
+/*   Updated: 2020/01/22 17:08:01 by asultanb         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/ft_ls.h"
 
-int     is_dir(char *arg)
+int		is_dir(char *arg)
 {
-    t_stat  checkdir;
+	t_stat	checkdir;
 
-    if (lstat(arg, &checkdir) == -1)
-        return (0);
-    if (!(S_ISDIR(checkdir.st_mode)))
-        return (0);
-    return (1);
+	if (lstat(arg, &checkdir) == -1)
+		return (0);
+	if (!(S_ISDIR(checkdir.st_mode)))
+		return (0);
+	return (1);
 }
 
-int     is_file(char *arg)
+int		is_file(char *arg)
 {
-    t_stat  checkfile;
+	t_stat	checkfile;
 
-    if (lstat(arg, &checkfile) == -1)
-        return (-1); 
-    if (!(S_ISDIR(checkfile.st_mode)))
-        return (1);
-    return (0);
+	if (lstat(arg, &checkfile) == -1)
+		return (-1);
+	if (!(S_ISDIR(checkfile.st_mode)))
+		return (1);
+	return (0);
 }
 
-int		check_dir(unsigned int flags, t_dirent *dp)
+int		check_dir(uint8_t flags, t_dirent *dp)
 {
 	if (flags & A_UPP && !(flags & A_LOW))
 	{
